Split tcp_channel/lib.rs
This change splits the tcp_channel crate into a few more submodules to
make the code a bit more wieldly. In particular, the two structs of
relevance, TcpReceiver and TcpSender, get to reside in their own file
each.
